// client/src/types.ts
export type ColorBlindnessType = 'none' | 'redgreen' | 'blueyellow' | 'grayscale';

/**
 * Interface representing an accessibility setting, which contains:
 * - colorBlindness - The type of colorblindness an account user has. Optional field.
 * - lowVision - Indicates whether the account user has low vision needs. Optional field.
 */
export interface AccessibilitySetting {
  colorBlindness?: ColorBlindnessType;
  lowVision?: boolean;
}